requirements.txt
bittensor==6.9.3
diffusers==0.28.0
transformers==4.40.0
accelerate==0.27.2
cryptography==42.0.0
ImageHash==4.3.1
loguru==0.7.0
Pillow==10.2.0
pydantic==1.10.13
PyYAML==6.0.1
Requests==2.31.0
setuptools==68.0.0
slowapi==0.1.8
tqdm==4.65.0
omegaconf==2.3.0
gitpython==3.1.40
httpx==0.26.0
ray[serve]==2.23.0
git+https://github.com/toilaluan/controlnet_aux.git
timm==1.0.3
numpy==1.26.4
pandas
typesense
clean-text[gpl]
jinja2
insightface
onnxruntime-gpu
openai
sentencepiece
redis==5.0.6
ctranslate2==4.3.1
ninja=1.10.2
